+++ src/SSWR/AVIRead/AVIRTimedCaptureForm.cpp
@@ -184,7 +184,7 @@ void __stdcall SSWR::AVIRead::AVIRTimedCaptureForm::OnStartClicked(void *userObj
 			UInt32 denorm;
 			UOSInt frameSize;
 			me->currCapture->GetVideoInfo(&me->videoInfo, &norm, &denorm, &frameSize);
-			me->lastSaveTime = -me->interval;
+			me->lastSaveTime = (UInt32)-(Int32)me->interval;
 			me->frameCnt = 0;
 			me->saveCnt = 0;
 			Media::ColorProfile dProfile(Media::ColorProfile::CPT_SRGB);
